Refactor model imports and restructure shared models
