+++ frontend/libs/portal/model/FilterSet.ts
@@ -2,27 +2,4 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-export class Filter {
-  public name = '';
-  public values: string[] = [];
-
-  public constructor(name: string, values: string[]) {
-    this.name = name;
-    this.values = values;
-  }
-}
-
-export class FilterSetDto {
-  public _key = '';
-  public name = '';
-  public includedFilters: Filter[] = [];
-  public excludedFilters: Filter[] = [];
-  public tableName = ''; // do I need this here?
-}
-
-export class FilterSetRequestDto {
-  public name = '';
-  public includedFilters: Filter[] = [];
-  public excludedFilters: Filter[] = [];
-  public tableName = '';
-}
+export * from '@shared/model/FilterSet';
