+++ frontend/libs/portal/model/IObligation.ts
@@ -2,78 +2,5 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-export interface IObligation {
-  _key: string;
-  name: string;
-  nameDe: string;
-  type: string;
-  warnLevel: string;
-  description: string;
-  descriptionDe: string;
-  autoApproved: boolean;
-  created: string;
-  updated: string;
-  spdxid: string;
-  remark: string;
-}
-
-export interface IObligationResponseAll {
-  count: number;
-  items: IObligation[];
-}
-
-export interface IDefaultSelectItem {
-  text: string;
-  value: string;
-}
-
-export interface ISelectItemWithCount extends IDefaultSelectItem {
-  count: number;
-}
-
-export class ObligationDTO implements IObligation {
-  _key: string;
-  autoApproved: boolean;
-  created: string;
-  description: string;
-  descriptionDe: string;
-  name: string;
-  nameDe: string;
-  type: string;
-  updated: string;
-  warnLevel: string;
-  spdxid: string;
-  remark: string;
-  public constructor() {
-    this._key = '';
-    this.autoApproved = false;
-    this.created = '';
-    this.description = '';
-    this.descriptionDe = '';
-    this.name = '';
-    this.nameDe = '';
-    this.type = '';
-    this.updated = '';
-    this.warnLevel = '';
-    this.spdxid = '';
-    this.remark = '';
-  }
-}
-
-export default class Obligation extends ObligationDTO {
-  constructor(dto: ObligationDTO) {
-    super();
-    Object.assign(this, dto);
-  }
-}
-
-export const DEFAULT_CLASSIFICATION_NAMES = new Set<string>([
-  'Permissive',
-  'Copyleft (weak)',
-  'Copyleft (strong)',
-  'Copyleft (network protective)',
-  'Anti-Tivoization',
-  'Severe patent retaliation',
-  'Doing Business with US',
-  'Unclear or Ambiguous',
-]);
+export * from '@shared/model/IObligation';
+export {default} from '@shared/model/IObligation';
