+++ frontend/libs/shared/components/sidebar/NavItem.vue
@@ -42,7 +42,7 @@
 </template>
 
 <script setup lang="ts">
-import INavItem from '@disclosure-portal/model/INavItem';
+import INavItem from '@shared/model/INavItem';
 import {TOOLTIP_OPEN_DELAY_IN_MS} from '@shared/utils/constant';
 import {ref} from 'vue';
 import {useI18n} from 'vue-i18n';
