+++ frontend/libs/shared/model/Callback.ts
