+++ frontend/libs/shared/model/PolicyRule.ts
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 Mercedes-Benz Group AG and Mercedes-Benz AG
+//
+// SPDX-License-Identifier: Apache-2.0
+
 export interface License {
   key: string;
   identifier: string;
@@ -16,19 +20,16 @@ export interface PolicyRule {
   licenses: License[];
 }
 
-// Canonical policy categories applied to licenses
 export type PolicyCat = 'allow' | 'warn' | 'deny';
 
-// Definition for policy status filter buttons (UI metadata)
 export interface PolicyStatusDef {
   key: 'all' | PolicyCat;
   icon: string;
   outlineIcon?: string;
-  color: string; // Vuetify theme color name
-  labelKey: string; // i18n translation key
+  color: string;
+  labelKey: string;
 }
 
-// Shared status button definitions for reuse in dialogs/components
 export const POLICY_STATUS_DEFS: PolicyStatusDef[] = [
   {key: 'all', icon: 'mdi-format-list-bulleted', color: 'primary', labelKey: 'FILTER_ALL'},
   {
@@ -48,7 +49,6 @@ export const POLICY_STATUS_DEFS: PolicyStatusDef[] = [
   },
 ];
 
-// CLI-side PolicyState (duplicated to avoid portal model dependency)
 export enum PolicyState {
   ALLOW = 'allow',
   DENY = 'deny',
@@ -68,20 +68,24 @@ export const PolicyStates: PolicyState[] = [
   PolicyState.ALLOW,
 ];
 
-/**
- * Compare function for policy status to enable proper sorting
- * Order: DENY (0) < NOASSERTION (1) < QUESTIONED (2) < WARN (3) < NOT_SET (4)  < ALLOW (5)
- */
+export class PolicyRulesAssignmentsDto {
+  public status = '';
+  public key = '';
+  public name = '';
+  public description = '';
+  public type = '';
+}
+
 export function comparePolicyStatus(a: string, b: string): number {
   const statusWeight: Record<string, number> = {
     deny: 0,
     noassertion: 1,
     questioned: 2,
     warn: 3,
-    notSet: 4,
+    notset: 4,
     allow: 5,
   };
-  const aWeight = statusWeight[a?.toLowerCase()] ?? statusWeight['notSet'];
-  const bWeight = statusWeight[b?.toLowerCase()] ?? statusWeight['notSet'];
+  const aWeight = statusWeight[a?.toLowerCase()] ?? statusWeight.notset;
+  const bWeight = statusWeight[b?.toLowerCase()] ?? statusWeight.notset;
   return aWeight - bWeight;
 }
