+++ frontend/libs/shared/model/SchemaPostRequest.ts
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 2025 Mercedes-Benz Group AG and Mercedes-Benz AG
+//
+// SPDX-License-Identifier: Apache-2.0
+
+export default class SchemaPostRequest {
+  public name = '';
+  public version = '';
+  public description = '';
+  public label = '';
+}
